debugging

+++ models/finetune.py
@@ -50,7 +50,7 @@ def get_ft(train_dataset, config, device, prompt_template="a photo of [attr] [ob
     )
     if config.amp:
         ft, optimizer = amp.initialize(
-            ft, optimizer, opt_level="O2",
+            ft.float(), optimizer, opt_level="O2",
             keep_batchnorm_fp32=True, loss_scale="dynamic"
         )
